refactoor PipelineStageBase and ManualPipelineStage
